backend/pyproject.toml
[project]
name = "knowledge-graph-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "litellm>=1.79.0",
    "openai>=1.51.2",
    "mineru[all]>=2.5.1",
    "neo4j>=6.0.2",
    "numpy>=1.24.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-stubs==2.3.2.250926",
    "pydantic>=2.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
    "uvicorn>=0.24.0",
    "xlrd>=2.0.2",
]
